pipeline: disable SEEK ingestion
