--- ats-companies/seek.csv
@@ -1,9 +1,0 @@
-name,slug,url
-JobStreet Indonesia,id,https://id.jobstreet.com
-JobStreet Malaysia,my,https://my.jobstreet.com
-JobStreet Philippines,ph,https://ph.jobstreet.com
-JobStreet Singapore,sg,https://sg.jobstreet.com
-JobsDB Hong Kong,hk,https://hk.jobsdb.com
-JobsDB Thailand,th,https://th.jobsdb.com
-SEEK Australia,au,https://www.seek.com.au
-SEEK New Zealand,nz,https://www.seek.co.nz
